se cambia el dir de alojamiento de la lib
Además se agrega el namespace Kumbia\ActiveRecord a todas las clases.
