+++ lib/Kumbia/ActiveRecord/Query/mysql_limit.php
@@ -19,18 +19,27 @@
  * @license    http://wiki.kumbiaphp.com/Licencia     New BSD License
  */
 
-namespace ActiveRecord\Query;
+namespace Kumbia\ActiveRecord\Query;
 
 /**
- * Obtiene el último id generado en mysql
+ * Adiciona limit y offset a la consulta sql en mysql
  *
- * @param PDO $dbh conexion pdo
- * @param string $pk campo clave primaria
- * @param string $table nombre de tabla
- * @param string $schema esquema
- * @return int
+ * @param string $sql consulta select
+ * @param string $limit valor limit
+ * @param string $offset valor offset
+ * @return string
  */
-function mysql_last_insert_id($dbh, $pk, $table, $schema = null)
+function mysql_limit($sql, $limit = null, $offset = null)
 {
-    return $dbh->lastInsertId();
+    if ($limit !== null) {
+        $limit = (int) $limit;
+        $sql .= " LIMIT $limit";
+    }
+
+    if ($offset !== null) {
+        $offset = (int) $offset;
+        $sql .= " OFFSET $offset";
+    }
+
+    return $sql;
 }
