+++ lib/Kumbia/ActiveRecord/Query/pgsql_last_insert_id.php
@@ -19,7 +19,7 @@
  * @license    http://wiki.kumbiaphp.com/Licencia     New BSD License
  */
 
-namespace ActiveRecord\Query;
+namespace Kumbia\ActiveRecord\Query;
 
 /**
  * Obtiene el último id generado en pgsql
